requirements.txt
opencv-python-headless
numpy
mss
pyobjc-framework-Quartz
pyautogui
ffmpeg-python
Pillow
imagehash 